finishing up corrections in push req #7

+++ Noted/NTDPagingCollectionViewLayout.h
@@ -12,11 +12,11 @@
 
 @property (nonatomic, assign) CGFloat pannedCardXTranslation;
 @property (nonatomic) int activeCardIndex;
-@property (nonatomic, readonly) float currentOptionsOffset;
+@property (nonatomic, readonly) CGFloat currentOptionsOffset;
 
 @property (nonatomic, strong) UIView *settingsView;
 
-- (void) finishAnimationWithVelocity:(float)velocity completion:(void (^)(void))completionBlock ;
-- (void) revealOptionsViewWithOffset:(float)offset;
-- (void) hideOptionsWithVelocity:(float)velocity completion:(void (^)(void))completionBlock;
+- (void) finishAnimationWithVelocity:(CGFloat)velocity completion:(void (^)(void))completionBlock ;
+- (void) revealOptionsViewWithOffset:(CGFloat)offset;
+- (void) hideOptionsWithVelocity:(CGFloat)velocity completion:(void (^)(void))completionBlock;
 @end
